requirements.txt
# requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
python-dotenv==1.0.1
pymongo==4.7.2
motor==3.4.0
bcrypt==4.1.2
pydub==0.25.1
yt-dlp==2024.04.09
eventlet==0.35.2
gunicorn==22.0.0  # <-- Add this back